Prefer salted MD5 instead of bcrypt
While Wesnoth 1.14 is slated to receive support for bcrypt in the
upcoming weeks, it's not feasible at this point to update 1.12 clients
with the required code, so we'll just have to keep using salted MD5 for
a while until 1.12's player base is sufficiently small at some point
after 1.14.0 is released. Once we switch to the upstream default of
bcrypt, all old clients will become permanently unable to log into the
MP server, so this is a rather crucial tradeoff in the interest of
backwards compatibility.

See wesnoth/wesnoth#1761 for additional information.