Prefer bcrypt over argon2
Wesnoth 1.14 supports bcrypt and MD5-hashed passwords, and we can't keep
playing catch with upstream all the time due to the nature of the client
and server interaction: the client needs to hash the password itself,
and clients are guaranteed to drift behind either Wesnoth or phpBB's
upstreams version-wise.

For the foreseeable future and to avoid issues, we are going to stick
with bcrypt instead of argon2. Wesnoth 1.12 does not support bcrypt, but
we don't need to worry about it anymore -- the userbase is minimal or
nonexistent by this point, and the official 1.12 server does not mandate
authentication like the 1.14 server does.

See wesnoth/wesnoth#1761 for additional information.

(I hate having to keep passing variations of this commit around in 2020.
The previous version was 6fceae5768acfcdd0ca56d4049a62594d5e4d309.)