Disable update_hashes cron task in migration
Wesnoth 1.14 and 1.15.x do not support combined hashes as of this
writing, meaning that hashes from players who are migrated automatically
(by hashing the existing MD5s with bcrypt) will be considered invalid by
the multiplayer server.

While the fix for that issue is relatively simple (players just need to
log in with their password into phpBB using the forums web proper so
phpBB hashes the password with bcrypt instead of using the MD5s) it is
also very inconvenient for people and very likely to cause confusion.

For now, we'll just avoid migrating old hashes just like we did with
phpBB 3.2.x (although that was also to avoid the invalid MD5-on-MD5
combination) so people who do log in through the forums get upgraded to
bcrypt and nobody needs to do anything specific to continue
authenticating on the MP server.